T-014: Reorganize auth page with route group components
