+++ apps/web/app/(auth)/page.tsx
@@ -1,0 +1,5 @@
+import LoginScreen from "./_components/login/LoginScreen";
+
+export default function AuthEntryPage() {
+  return <LoginScreen />;
+}
