--- apps/web/app/page.tsx
@@ -1,5 +1,0 @@
-import LoginForm from "./login-form";
-
-export default function HomePage() {
-  return <LoginForm />;
-}
